auto-video/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-video"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.0.0",
    "elevenlabs>=1.0.0",
    "moviepy>=2.0.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[tool.setuptools.packages.find]
include = ["scripts*"]
